requirements.txt
fastapi
uvicorn
librosa
scikit-learn
numpy
joblib
soundfile
psycopg2-binary
python-multipart
pydantic
azure-storage-blob>=12.20.0
python-dotenv
